Move notifier package under internal
